Add files via upload
This project, undertaken on Datacamp.com, involves the implementation of code based on provided instructions. It explores the phyllotaxis phenomenon, where the arrangement of leaves on a plant stem follows a spiral pattern. Phyllotaxis serves as a compelling illustration of how mathematical principles can accurately describe patterns observed in nature. The primary objective of this project is to create simulated flowers by leveraging the principles of phyllotaxis.